git: Fuck (ai testmerge)

+++ Assets/Scripts/IStats.cs
@@ -1,4 +1,4 @@
-public interface IStats { 
-        public void ModifyHealth(byte modify); 
+public interface IStats {
+        public void ModifyHealth(int modify); 
         void Die();
 }
